Add middlewares and changes to config
