+++ api/database/__init__.py
@@ -2,10 +2,10 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from config import SQLALCHEMY_DATABASE_URI
+from config import Config
 
 # Create database engine
-engine = create_engine(SQLALCHEMY_DATABASE_URI, echo=False)
+engine = create_engine(Config.SQLALCHEMY_DATABASE_URI, echo=False)
 
 # Create database session
 Session = sessionmaker(bind=engine)
