+++ manage.py
@@ -1,3 +1,4 @@
-import app
+from app import app
 
-app.create_app()
+if __name__ == '__main__':
+    app.run()
